运行测试/未验证测试区/requirements.txt
# 核心依赖
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=0.21.0

# AI集成
openai>=1.0.0
tiktoken>=0.4.0

# 技术分析
ta-lib>=0.4.25

# 调度与异步
schedule>=1.2.0
asyncio-throttle>=1.0.2

# 数据处理
scipy>=1.9.0
scikit-learn>=1.1.0

# 日志与监控
colorlog>=6.7.0

# 开发工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
ruff>=0.0.280

# 类型检查
mypy>=1.0.0
types-requests>=2.28.0